to-do: rip the inspector out
